+++ src/_util_rowwise.jl
@@ -1,9 +1,26 @@
 """
-    rowmean(X)
-Compute row-wise means of a matrix.
+    rowsum(X)
+Row-wise sums of a matrix.
 * `X` : Data (n, p).
 
-Return a vector.
+Return a vector (n).
+
+## Examples
+```julia
+using Jchemo
+ 
+X = rand(5, 2) 
+rowsum(X)
+```
+""" 
+rowsum(X) = colsum(ensure_mat(X)')
+
+"""
+    rowmean(X)
+Row-wise means of a matrix.
+* `X` : Data (n, p).
+
+Return a vector (n).
 
 ## Examples
 ```julia
@@ -18,16 +35,13 @@ rowmean(X) = colmean(ensure_mat(X)')
 
 """
     rownorm(X)
-Compute row-wise norms of a matrix.
+Row-wise norms of a matrix.
 * `X` : Data (n, p).
+
+Return a vector (n).
 
 The norm of each row x of `X` is computed as:
 * sqrt(x' * x)
-
-Return a vector.
-
-Note: Thanks to @mcabbott 
-at https://discourse.julialang.org/t/orders-of-magnitude-runtime-difference-in-row-wise-norm/96363.
 
 ## Examples
 ```julia
@@ -39,14 +53,35 @@ X = rand(n, p)
 rownorm(X)
 ```
 """ 
-rownorm(X) = colnorm(ensure_mat(X)')
+rownorm(X) = sqrt.(rownorm2(X))
+
+"""
+    rownorm2(X)
+Row-wise squared norms of a matrix.
+* `X` : Data (n, p).
+
+Return a vector (n).
+
+See function `rownorm`.
+
+## Examples
+```julia
+using Jchemo
+
+n, p = 5, 6
+X = rand(n, p)
+
+rownorm2(X)
+```
+""" 
+rownorm2(X) = colnorm2(ensure_mat(X)')
 
 """
     rowstd(X)
-Compute row-wise standard deviations (uncorrected) of a matrix`.
+Row-wise (uncorrected) standard deviations of a matrix`.
 * `X` : Data (n, p).
 
-Return a vector.
+Return a vector (n).
 
 ## Examples
 ```julia
@@ -57,31 +92,14 @@ X = rand(n, p)
 rowstd(X)
 ```
 """ 
-rowstd(X) = colstd(ensure_mat(X)')
-
-"""
-    rowsum(X)
-Compute row-wise sums of a matrix.
-* `X` : Data (n, p).
-
-Return a vector.
-
-## Examples
-```julia
-using Jchemo
- 
-X = rand(5, 2) 
-rowsum(X)
-```
-""" 
-rowsum(X) = colsum(ensure_mat(X)')
+rowstd(X) = sqrt.(rowvar(X))
 
 """
     rowvar(X)
-Compute row-wise variances (uncorrected) of a matrix.
+Row-wise (uncorrected) variances of a matrix.
 * `X` : Data (n, p).
 
-Return a vector.
+Return a vector (n).
 
 ## Examples
 ```julia
@@ -92,11 +110,11 @@ X = rand(n, p)
 rowvar(X)
 ```
 """ 
-rowvar(X) = rowvar(ensure_mat(X)')
+rowvar(X) = colvar(ensure_mat(X)')
 
 ##### Functions skipping missing data
+rowsumskip(X) = colsumskip(ensure_mat(X)')
 rowmeanskip(X) = colmeanskip(ensure_mat(X)')
 rowstdskip(X) = colstdskip(ensure_mat(X)')
-rowsumskip(X) = colsumskip(ensure_mat(X)')
 rowvarskip(X) = colvarskip(ensure_mat(X)')